🐛 Fix & CI: Fix CI timeout and cache path.
🐛 Fix & CI: Fix CI timeout and cache path.

~/.conan -> $HOME/.conan

🐛 Fix & CI: Try to fix timeout problem.

Travis-CI build timeout is 50min. When we do a fresh build the timeout is easily reached because we need to build the dependencies. We need to terminate the build before the timeout so we can upload the cache and use it when we restart the build.

🐛 Fix & CI: Try to fix apt-get timeout.

CI lags at Processing triggers for libc-bin (2.27-3ubuntu1) ... when installing libjack-dev.

🐛 Fix & CI: Can't figure why apt-get is lagging on man-db.

⚡️ CI: Same.

⚡️ CI: Same.

⚡️ CI: Same.

⚡️ CI: Same.