⚡️ Perf: Video decoding&rendering with double buffer
In the past we use a list of `VideoFrame` buffer
in the `AsyncVideoProvider` for both video display and other
place using the video frames. However, AVP is not singleton
and such flow involves multithread calling to VideoDisplay::Render
which is not safe.

Also, this fixes a bug: On macOS when the video is playing, using
the visual tool(move mouse on the video box) will cause the video
to lag because mouse motion causes a lot of events in the evt queue
and triggers Render() in a high frequent. And the FrameReadyEvent
fired by AVP is messed up with them. However those mouse events don't
contain new frame data, so the display just repaints with old frame
data, and new frame data is delayed to upload to video mem.

Now we use a `wxTimer` to force the display to render at fixed rate,
And every time it renders, the latest frame data is gather from AVP.